Sync as of 11:12AM PST 11/01/2001

+++ server/mp4creator/mp4creator.h
@@ -54,7 +54,7 @@
 #endif
 
 MP4CREATOR_GLOBAL char* ProgName;
-MP4CREATOR_GLOBAL u_int32_t VideoFrameRate;
+MP4CREATOR_GLOBAL float VideoFrameRate;
 MP4CREATOR_GLOBAL u_int32_t MaxPayloadSize;
 
 #endif /* __MP4CREATOR_INCLUDED__ */
